Support long numbers in 2024 answers

+++ 2024/src/main/kotlin/Main.kt
@@ -18,7 +18,7 @@ enum class Colors {
     }
 }
 
-fun printAnswer(day: Int, part: Int, answer: Int) {
+fun printAnswer(day: Int, part: Int, answer: Long) {
     val formattedOutput = "%,d".format(answer)
     val output = if (formattedOutput != "$answer") "\t($formattedOutput)" else ""
     println("${Colors.BLUE}[Day $day]${Colors.CYAN} Part $part:${Colors.RESET} $answer$output")
